requirements.txt
# python -m build && python -m twine upload --repository testpypi dist/*
# python -m pip install --index-url https://test.pypi.org/simple/ --no-deps yolov5test4
# python -c "from sample import simple; print(simple.add_one(10))"

# Example requirements
numpy>=1.18.5

# PyPI package deployment
build
twine

requests

pick
gitpython